pyproject.toml
[tool.poetry]
authors = ["Sylvain Boissel <sylvain.boissel@dgcl.gouv.fr>"]
description = "Open Collectivités is a web portal aiming at giving an easy access to statistical data about French local collectivities."
license = "AGPL-3"
name = "opencollectivites"
version = "1.0.0"

[tool.poetry.dependencies]
Babel = "^2.9.1"
Django = "^3.2.7"
Pillow = "^8.3.2"
Pygments = "^2.10.0"
Unidecode = "^1.2.0"
beautifulsoup4 = "^4.9.3"
dateparser = "^1.0.0"
django-analytical = "^3.0.0"
django-cors-headers = "^3.8.0"
django-dsfr = "^0.5.0"
django-francedata = "^0.14.3"
django-json-widget = "^1.1.1"
django-ninja = "^0.14.0"
django-simple-history = "^3.0.0"
django-widget-tweaks = "^1.4.8"
feedparser = "^6.0.8"
gunicorn = "^20.1.0"
openpyxl-dictreader = "^0.1.3"
psycopg2-binary = "^2.9.1"
pyRFC3339 = "^1.1"
python = "^3.8"
python-stdnum = "^1.16"
requests = "^2.26.0"
wagtail-markdown = "^0.7.0"
wagtailsvg = "^0.0.20"
xlrd = "^2.0.1"
xmltodict = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
coverage = "^5.5"
mock = "^4.0.3"
pylint = "^2.10.2"
requests-mock = "^1.9.3"
responses = "^0.14.0"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
